refactor: reorganize component file structure and update analysis mock data for tests
