+++ src/ravioli/frontend/src/components/analysis/Insights.ts
