+++ src/ravioli/frontend/src/components/analysis/create-analysis/templates.ts
@@ -1,4 +1,4 @@
-import { store } from '../../store';
+import { store } from '../../../store';
 
 export function renderSelection() {
     return `
