+++ src/ravioli/frontend/src/components/analysis/insights/interactions.ts
@@ -1,6 +1,6 @@
-import { api } from '../../services/api';
+import { api } from '../../../services/api';
 import { format } from 'date-fns';
-import type { LineageNode, LineageEdge, LineageResponse } from '../../types';
+import type { LineageNode, LineageEdge, LineageResponse } from '../../../types';
 import { state } from './state';
 import { banCard, insightPill, renderNodeElement, renderExpandButton } from './templates';
