+++ src/ravioli/frontend/src/components/analysis/notebook/templates.ts
