+++ src/ravioli/frontend/src/main.ts
@@ -3,8 +3,8 @@ import { store } from './store';
 import { api } from './services/api';
 import { renderSidebar, updateSidebarUI } from './components/Sidebar';
 import { renderAnalysis, updateAnalysisUI } from './components/analysis/AnalysisShell';
-import { renderInsights } from './components/Insights';
-import { renderCreateAnalysis } from './components/CreateAnalysis';
+import { renderInsights } from './components/analysis/Insights';
+import { renderCreateAnalysis } from './components/analysis/CreateAnalysis';
 import { renderKnowledge } from './components/Knowledge';
 import { renderData } from './components/Data';
 import { renderSettings } from './components/Settings';
